separate implementations from base classes
