+++ xedocs/schemas/corrections/implementations/avg_se_gain.py
@@ -9,8 +9,8 @@ wiki:
 
 import rframe
 
-from .base_corrections import TimeSampledCorrection
-from ..constants import PARTITION
+from ..base_corrections import TimeSampledCorrection
+from ...constants import PARTITION
 
 
 class AvgSingleElectronGain(TimeSampledCorrection):
