+++ xedocs/schemas/corrections/implementations/baseline_samples_nv.py
@@ -8,7 +8,7 @@ wiki:
 """
 
 
-from .base_corrections import TimeSampledCorrection
+from ..base_corrections import TimeSampledCorrection
 
 
 class BaselineSamplesNV(TimeSampledCorrection):
