+++ xedocs/schemas/corrections/implementations/bayes_model.py
@@ -5,7 +5,7 @@ import datetime
 from typing import Literal
 import rframe
 
-from .base_references import BaseResourceReference
+from ..base_references import BaseResourceReference
 
 
 class NaiveBayesClassifier(BaseResourceReference):
