+++ xedocs/schemas/corrections/implementations/electron_diffusion_cte.py
@@ -7,7 +7,7 @@
 
 import rframe
 
-from .base_corrections import TimeSampledCorrection
+from ..base_corrections import TimeSampledCorrection
 
 
 class ElectronDiffusionCte(TimeSampledCorrection):
