+++ xedocs/schemas/corrections/implementations/photoionization_strength.py
@@ -3,7 +3,7 @@
 https://xe1t-wiki.lngs.infn.it/doku.php?id=xenon:xenonnt_sr1:photoionization_origin
 """
 
-from .base_corrections import TimeSampledCorrection
+from ..base_corrections import TimeSampledCorrection
 
 
 class PhotoionizationStrength(TimeSampledCorrection):
