+++ xedocs/schemas/corrections/implementations/pmt_area_to_pe.py
@@ -8,8 +8,8 @@
 
 import rframe
 
-from .base_corrections import TimeSampledCorrection
-from ..constants import DETECTOR
+from ..base_corrections import TimeSampledCorrection
+from ...constants import DETECTOR
 
 
 class PmtAreaToPE(TimeSampledCorrection):
