+++ xedocs/schemas/corrections/implementations/som_network.py
@@ -4,7 +4,7 @@ Self-Organizing Map to classify peaklets as one of several SOM_types,
 as well as classifying those types into either an S1 or an S2
 """
 
-from .base_references import BaseResourceReference
+from ..base_references import BaseResourceReference
 
 
 class SOMClassifier(BaseResourceReference):
