+++ xedocs/schemas/corrections/implementations/z_bias.py
@@ -5,7 +5,7 @@ the observed r(r_obs) and z(z_obs).
 Reference: xenon:xenonnt:terliuk:drift_field_z_bias_correction
 """
 
-from .base_references import BaseResourceReference
+from ..base_references import BaseResourceReference
 
 class ZBias(BaseResourceReference):
     _ALIAS = 'z_bias_maps'
